src/main/java/com/vojik/solutions/Solution_31.java
package com.vojik.solutions;

/**
 * 31. Next Permutation
 */
public class Solution_31 {

  public void nextPermutation(int[] nums) {
    int len = nums.length;
    if (len == 1) {
      return;
    }
    int i = len - 2;
    while (i >= 0 && nums[i] >= nums[i + 1]) {
      i--;
    }

    if (i >= 0) {
      int j = nums.length - 1;
      while (nums[j] <= nums[i]) {
        j--;
      }
      swap(nums, i, j);
    }
    reverse(nums, i + 1);
  }

  private void reverse(int[] nums, int i) {
    int j = nums.length - 1;
    while (i < j) {
      swap(nums, i, j);
      i++;
      j--;
    }
  }

  private void swap(int[] nums, int i, int j) {
    int temp = nums[i];
    nums[i] = nums[j];
    nums[j] = temp;
  }

}
